[Docs] String+ 파일로 메서드 이동
